requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
chromadb
faiss-cpu
langchain-community
langchain-google-genai

pypdf2

